主キーをserialに変更

+++ supabase/migrations/20260926100004_create_cart_and_favorite_bot.sql
@@ -3,13 +3,12 @@
 -- 同じユーザーが同じチャットボットを二重にカート・お気に入りに入れられないよう
 -- (user_id, chatbot_id) に UNIQUE 制約を付ける。
 --
--- 注意:
---   主キーは smallserial のため、採番できるのは 1〜32767 まで。
---   行を削除しても番号は再利用されないので、カートの出し入れが多いと
---   いずれ上限に達する。足りなくなったら integer / bigint に変更すること。
+-- 主キーは serial(integer) にする。
+--   smallserial だと採番できるのが 32767 までで、行を削除しても番号は
+--   再利用されないため、カートの出し入れやお気に入りの付け外しで上限に達しやすい。
 
 create table cart (
-  cart_id smallserial primary key,
+  cart_id serial primary key,
   chatbot_id uuid not null references chatbot(chatbot_id),
   user_id varchar(20) not null references users(user_id) on update cascade,
   create_at timestamp with time zone not null default now(),
@@ -17,7 +16,7 @@ create table cart (
 );
 
 create table favorite_bot (
-  favorite_bot_id smallserial primary key,
+  favorite_bot_id serial primary key,
   chatbot_id uuid not null references chatbot(chatbot_id),
   user_id varchar(20) not null references users(user_id) on update cascade,
   create_at timestamp with time zone not null default now(),
